requirements.txt
netaddr
Flask==0.11.1
flask_restful==0.3.5
flask_cors
requests
pyyaml
dpath